[FEAT] User 정보 커스텀 어노테이션 구현
