--- src/main/java/com/example/hackathon/auth/client/KakaoOAuthClient.java
@@ -1,60 +1,0 @@
-package com.example.hackathon.auth.client;
-
-import com.example.hackathon.auth.dto.KakaoUserInfoDto;
-import com.example.hackathon.auth.dto.TokenResponseDto;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.*;
-import org.springframework.stereotype.Component;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-
-@Component
-@RequiredArgsConstructor
-public class KakaoOAuthClient {
-
-    @Value("${oauth.kakao.client-id}")
-    private String clientId;
-
-    @Value("${oauth.kakao.redirect-uri}")
-    private String redirectUri;
-
-    private final RestTemplate restTemplate;
-
-    public TokenResponseDto getToken(String code) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add("grant_type", "authorization_code");
-        params.add("client_id", clientId);
-        params.add("redirect_uri", redirectUri);
-        params.add("code", code);
-
-        HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(params, headers);
-        ResponseEntity<TokenResponseDto> response = restTemplate.postForEntity(
-                "https://kauth.kakao.com/oauth/token",
-                request,
-                TokenResponseDto.class
-        );
-
-        return response.getBody();
-    }
-
-    public KakaoUserInfoDto getUserInfo(String accessToken) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setBearerAuth(accessToken);
-
-        HttpEntity<Void> request = new HttpEntity<>(headers);
-        ResponseEntity<KakaoUserInfoDto> response = restTemplate.exchange(
-                "https://kapi.kakao.com/v2/user/me",
-                HttpMethod.GET,
-                request,
-                KakaoUserInfoDto.class
-        );
-
-        return response.getBody();
-    }
-}
-
