--- src/main/java/com/example/hackathon/auth/dto/TokenResponseDto.java
@@ -1,23 +1,0 @@
-package com.example.hackathon.auth.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class TokenResponseDto {
-
-    @JsonProperty("access_token")
-    private String accessToken;
-    @JsonProperty("token_type")
-    private String tokenType;
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-    @JsonProperty("expires_in")
-    private Long expiresIn;
-
-}
-
