--- src/main/java/com/example/hackathon/auth/service/OAuthService.java
@@ -1,6 +1,0 @@
-package com.example.hackathon.auth.service;
-
-public interface OAuthService {
-    String loginWithKakao(String code);
-}
-
