--- src/main/java/com/example/hackathon/common/exception/CustomOAuthException.java
@@ -1,8 +1,0 @@
-package com.example.hackathon.common.exception;
-
-public class CustomOAuthException extends RuntimeException {
-    public CustomOAuthException(String message) {
-        super(message);
-    }
-}
-
