--- src/main/java/com/example/hackathon/common/exception/GlobalExceptionHandler.java
@@ -1,178 +1,0 @@
-package com.example.hackathon.common.exception;
-
-import com.example.hackathon.common.response.Code;
-import com.example.hackathon.common.response.Response;
-import jakarta.validation.ConstraintViolationException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.BeanCreationException;
-import org.springframework.beans.factory.UnsatisfiedDependencyException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.IncorrectResultSizeDataAccessException;
-import org.springframework.dao.InvalidDataAccessApiUsageException;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageConversionException;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.orm.jpa.JpaSystemException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.NoSuchElementException;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler(CustomOAuthException.class)
-    public ResponseEntity<String> handleOAuthException(CustomOAuthException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("OAuth 오류: " + e.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleGeneralException(Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("서버 오류: " + e.getMessage());
-    }
-
-    @ExceptionHandler(BusinessException.class)
-    public ResponseEntity<Response<Void>> businessExceptionHandler(BusinessException ex) {
-
-        return ResponseEntity
-                .status(ex.getReason().getStatus())
-                .body(Response.fail(ex.getCode()));
-    }
-
-    @ExceptionHandler(GlobalException.class)
-    public ResponseEntity<Response<Void>> globalExceptionHandler(GlobalException ex) {
-
-        return ResponseEntity
-                .status(ex.getReason().getStatus())
-                .body(Response.fail(ex.getCode()));
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatusCode status, WebRequest request) {
-
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getFieldErrors()
-                .forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, errors));
-    }
-
-    @Override
-    protected ResponseEntity<Object> handleHttpMessageNotReadable(HttpMessageNotReadableException ex, HttpHeaders headers, HttpStatusCode status, WebRequest request) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-
-    @ExceptionHandler(BeanCreationException.class)
-    public ResponseEntity<Object> handleBeanCreationException(BeanCreationException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(ClassCastException.class)
-    public ResponseEntity<Object> handleClassCastException(ClassCastException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<Object> handleConstraintViolationException(ConstraintViolationException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<Object> handleDataIntegrityViolationException(DataIntegrityViolationException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(HttpMessageConversionException.class)
-    public ResponseEntity<Object> handleHttpMessageConversionException(HttpMessageConversionException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<Object> handleIllegalArgumentException(IllegalArgumentException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(IncorrectResultSizeDataAccessException.class)
-    public ResponseEntity<Object> handleIncorrectResultSizeException(IncorrectResultSizeDataAccessException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(InvalidDataAccessApiUsageException.class)
-    public ResponseEntity<Object> handleInvalidDataAccessApiUsageException(InvalidDataAccessApiUsageException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(JpaSystemException.class)
-    public ResponseEntity<Object> handleJpaSystemException(JpaSystemException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(NoSuchElementException.class)
-    public ResponseEntity<Object> handleNoSuchElementException(NoSuchElementException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Response.fail(Code.BAD_REQUEST, ex.getMessage()));
-    }
-
-    @ExceptionHandler(NullPointerException.class)
-    public ResponseEntity<Object> handleNullPointerException(NullPointerException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(UnsatisfiedDependencyException.class)
-    public ResponseEntity<Object> handleUnsatisfiedDependencyException(UnsatisfiedDependencyException ex) {
-
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response.fail(Code.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-}
-
-
-
