--- src/main/java/com/example/hackathon/common/exception/UsernameNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.hackathon.common.exception;
-
-public class UsernameNotFoundException  extends RuntimeException {
-    public UsernameNotFoundException(String message) {
-        super(message);
-    }
-}
