--- src/main/java/com/example/hackathon/common/jwt/JwtAuthenticationFilter.java
@@ -1,61 +1,0 @@
-package com.example.hackathon.common.jwt;
-
-import com.example.hackathon.common.util.JwtUtil;
-import io.jsonwebtoken.JwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-
-    private final JwtUtil jwtUtil;
-//    private final UserRepository userRepository;
-
-    @Override
-    protected void doFilterInternal(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            FilterChain filterChain
-    ) throws ServletException, IOException {
-
-        String token = resolveToken(request);
-
-        try {
-            if (token != null && jwtUtil.validateToken(token)) {
-                String id = jwtUtil.extractKakaoId(token);
-
-//                User user = userRepository.findByEmail(email)
-//                        .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다: " + email));
-
-                // 인증 객체 생성 (권한 정보는 생략하거나 필요시 추가)
-                UsernamePasswordAuthenticationToken authentication =
-                        new UsernamePasswordAuthenticationToken(new Object(), null, null);
-
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        } catch (JwtException | IllegalArgumentException e) {
-            logger.warn("JWT 인증 실패: {}", e);
-        }
-
-        filterChain.doFilter(request, response);
-    }
-
-    private String resolveToken(HttpServletRequest request) {
-        String bearer = request.getHeader("Authorization");
-        if (bearer != null && bearer.startsWith("Bearer ")) {
-            return bearer.substring(7);
-        }
-        return null;
-    }
-}
-
