--- src/main/java/com/example/hackathon/common/response/BaseCode.java
@@ -1,5 +1,0 @@
-package com.example.hackathon.common.response;
-
-public interface BaseCode {
-    ReasonDTO getReason();
-}
