--- src/main/java/com/example/hackathon/common/response/ReasonDTO.java
@@ -1,14 +1,0 @@
-package com.example.hackathon.common.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDTO {
-
-    private final HttpStatus status;
-    private final String code;
-    private final String message;
-}
