+++ src/main/java/com/example/hackathon/global/auth/service/OAuthService.java
@@ -1,0 +1,6 @@
+package com.example.hackathon.global.auth.service;
+
+public interface OAuthService {
+    String loginWithKakao(String code);
+}
+
