+++ src/main/java/com/example/hackathon/global/config/WebConfig.java
@@ -1,0 +1,17 @@
+package com.example.hackathon.global.config;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addRedirectViewController("/swagger", "/swagger-ui/index.html");
+        registry.addRedirectViewController("/swagger/", "/swagger-ui/index.html");
+    }
+}
