+++ src/main/java/com/example/hackathon/global/exception/CustomOAuthException.java
@@ -1,0 +1,8 @@
+package com.example.hackathon.global.exception;
+
+public class CustomOAuthException extends RuntimeException {
+    public CustomOAuthException(String message) {
+        super(message);
+    }
+}
+
