+++ src/main/java/com/example/hackathon/global/exception/UsernameNotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.hackathon.global.exception;
+
+public class UsernameNotFoundException  extends RuntimeException {
+    public UsernameNotFoundException(String message) {
+        super(message);
+    }
+}
