+++ src/main/java/com/example/hackathon/global/response/BaseCode.java
@@ -1,0 +1,5 @@
+package com.example.hackathon.global.response;
+
+public interface BaseCode {
+    ReasonDTO getReason();
+}
