+++ src/main/java/com/example/hackathon/global/response/ReasonDTO.java
@@ -1,0 +1,14 @@
+package com.example.hackathon.global.response;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ReasonDTO {
+
+    private final HttpStatus status;
+    private final String code;
+    private final String message;
+}
